Ptrace Architecture Dependent
